third_party/WebKit/LayoutTests/css3/motion-path/offsetParent.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<section id="container" style="will-change: transform">
    <div id="d1" style="offset-position: 0% 0%">
        <span id="s1"></span>
    </div>
    <div id="d2" style="offset-path: path('m 0 0 z')">
        <span id="s2"></span>
    </div>
   <div id="d3" style="offset-distance: 100%">
        <span id="s3"></span>
    </div>
   <div id="d4" style="offset-rotate: 360deg">
        <span id="s4"></span>
    </div>
   <div id="d5" style="offset-anchor: bottom right">
        <span id="s5"></span>
    </div>
</section>
<script>
'use strict';
test(function() {
    assert_equals(s1.offsetParent, d1);
}, "inside offset-position");
test(function() {
    assert_equals(s2.offsetParent, d2);
}, "inside offset-path");
test(function() {
    assert_equals(s3.offsetParent, container);
}, "inside offset-distance");
test(function() {
    assert_equals(s4.offsetParent, container);
}, "inside offset-rotate");
test(function() {
    assert_equals(s5.offsetParent, container);
}, "inside offset-anchor");
</script>
